Move files
Move data files and data processing files into respective folders
Also, add dynamic file location into the scripts
