+++ .gitignore
@@ -1,2 +1,3 @@
 samples/
 **/__pycache__/
+data_files/
