+++ data_processing/description.py
@@ -1,7 +1,26 @@
 import json
+import os
+
+if os.name == 'nt':
+    current_directory = os.path.dirname(os.path.realpath(__file__)) # Get current directory
+else:
+    current_directory = os.path.dirname(os.path.realpath(__name__)) # Get current directory
+    
+
+path = current_directory.split(os.sep)
+
+root_index = path.index('Capstone-Team14')
+root_dir = os.sep.join(path[:root_index+1])
+data_dir = os.path.join(root_dir, 'data_files', 'uploaded_schedule')
+
+try:
+    os.makedirs(data_dir)
+except:
+    pass
 
 # Load the JSON file
-file_path = "filtered.json"
+
+file_path = os.path.join(data_dir, "filtered.json")
 with open(file_path, "r") as file:
     data = json.load(file)
