+++ data_processing/processor.py
@@ -1,4 +1,5 @@
 import json
+import os
 
 def get_cist_csci_courses(json_file_path):
     with open(json_file_path, 'r', encoding='utf-8') as f:
@@ -16,7 +17,19 @@ def get_cist_csci_courses(json_file_path):
 
     return filtered_courses
 if __name__ == "__main__":
-    json_file = "fa21-fa24.json"
+    if os.name == 'nt':
+        current_directory = os.path.dirname(os.path.realpath(__file__)) # Get current directory
+    else:
+        current_directory = os.path.dirname(os.path.realpath(__name__)) # Get current directory
+        
+
+    path = current_directory.split(os.sep)
+
+    root_index = path.index('Capstone-Team14')
+    root_dir = os.sep.join(path[:root_index+1])
+    data_dir = os.path.join(root_dir, 'data_files', 'uploaded_schedule')
+    json_file =  os.path.join(data_dir, "fa21-fa24.json")
+    
     result = get_cist_csci_courses(json_file)
     with open("filtered.json", "w", encoding='utf-8') as out:
         json.dump(result, out, indent=2)
