+++ docs/html/_formulas.log
@@ -1,4 +1,4 @@
-This is pdfTeX, Version 3.141592653-2.6-1.40.26 (TeX Live 2024) (preloaded format=latex 2025.3.4)  1 APR 2025 20:38
+This is pdfTeX, Version 3.141592653-2.6-1.40.26 (TeX Live 2024) (preloaded format=latex 2025.3.4)  1 APR 2025 20:41
 entering extended mode
  restricted \write18 enabled.
  %&-line parsing enabled.
