+++ docs/html/_formulas.tex
@@ -28,10 +28,10 @@
 \fi
 \pagestyle{empty}
 \begin{document}
-a \equiv c \Leftarrow a \equiv b, b \equiv c
+$ a \equiv c \Leftarrow a \equiv b, b \equiv c $
 \pagebreak
 
-b \in S \Leftarrow a \in S, a \equiv b
+$ b \in S \Leftarrow a \in S, a \equiv b $
 \pagebreak
 
 \end{document}
