+++ docs/html/namespaceprocess__constraints.html
@@ -225,9 +225,7 @@ Variables</h2></td></tr>
 </div><div class="memdoc">
 
 <p>Take formated constraints and convert it into asp rules In addition to the rules defined by the constraints parameter, the rules file is also appended with equivalence rules. </p>
-<ul>
-<li><picture><source srcset="form_0_dark.png" media="(prefers-color-scheme: dark)"/><img class="formulaInl" alt="a \equiv c \Leftarrow a \equiv b, b \equiv c" src="form_0.png"/></picture></li>
-<li><picture><source srcset="form_1_dark.png" media="(prefers-color-scheme: dark)"/><img class="formulaInl" alt="b \in S \Leftarrow a \in S, a \equiv b" src="form_1.png"/></picture></li>
+<p><picture><source srcset="form_0_dark.png" media="(prefers-color-scheme: dark)"/><img class="formulaInl" alt="$ a \equiv c \Leftarrow a \equiv b, b \equiv c $" src="form_0.png"/></picture> <picture><source srcset="form_1_dark.png" media="(prefers-color-scheme: dark)"/><img class="formulaInl" alt="$ b \in S \Leftarrow a \in S, a \equiv b $" src="form_1.png"/></picture></p><ul>
 <li>etc.</li>
 </ul>
 <dl class="params"><dt>Parameters</dt><dd>
