+++ docs/latex/namespaceprocess__constraints.tex
@@ -103,10 +103,8 @@ Definition at line \mbox{\hyperlink{process__constraints_8py_source_l00132}{132}
 
 Take formated constraints and convert it into asp rules In addition to the rules defined by the constraints parameter, the rules file is also appended with equivalence rules. 
 
-
+$ a \equiv c \Leftarrow a \equiv b, b \equiv c $ $ b \in S \Leftarrow a \in S, a \equiv b $
 \begin{DoxyItemize}
-\item a \equiv c \Leftarrow a \equiv b, b \equiv c
-\item b \in S \Leftarrow a \in S, a \equiv b
 \item etc.
 \end{DoxyItemize}
