+++ docs/man/man3/process_constraints.3
@@ -118,11 +118,7 @@ Definition at line \fB132\fP of file \fBprocess_constraints\&.py\fP\&.
 .SS " None process_constraints\&.convert_asp (dict constraints, str file)"
 
 .PP
-Take formated constraints and convert it into asp rules In addition to the rules defined by the constraints parameter, the rules file is also appended with equivalence rules\&. 
-.IP "\(bu" 2
-a \equiv c \Leftarrow a \equiv b, b \equiv c
-.IP "\(bu" 2
-b \in S \Leftarrow a \in S, a \equiv b
+Take formated constraints and convert it into asp rules In addition to the rules defined by the constraints parameter, the rules file is also appended with equivalence rules\&. $ a \equiv c \Leftarrow a \equiv b, b \equiv c $ $ b \in S \Leftarrow a \in S, a \equiv b $
 .IP "\(bu" 2
 etc\&.
 .PP
