+++ identifier.py
@@ -1,12 +1,23 @@
 import sys, clingo, math
-from jsonconverter import convert
+from data_processing.jsonconverter import convert
 import argparse
+import os
+
+if os.name == 'nt':
+    current_directory = os.path.dirname(os.path.realpath(__file__)) # Get current directory
+else:
+    current_directory = os.path.dirname(os.path.realpath(__name__)) # Get current directory
+path = current_directory.split(os.sep)
+
+root_index = path.index('Capstone-Team14')
+root_dir = os.sep.join(path[:root_index+1])
+read_dir = os.path.join(root_dir, 'data_files', 'uploaded_schedule')
 
 parser = argparse.ArgumentParser(description="Input class json")
 
 # Add arguments with default values
 parser.add_argument(
-    "--file", type=str, default="filtered.json", help="File name to process"
+    "--file", type=str, default=os.path.join(read_dir, "filtered.json"), help="File name to process"
 )
